[BE] refactor : auth/exception 폴더 생성
